Rename TypeScript storageService to fileStorageService
