+++ backend/typescript/services/interfaces/fileStorageService.ts
